Implemented preliminary polyphony support
Yay! Finally we have a MIDI sequencer that supports polyphony. And a
mixer plugin that combines the voices. Lots of changes were needed here,
and I'll be the first to admit that not every part of this is as elegant
as it could (and should!) be, e.g. hard-coded number of voices that must
match the MIDI file, hard-coded tempo (and no tempo changes), etc.

That we can fix, though.

Signed-off-by: Vegard Nossum <vegard.nossum@gmail.com>